going well

+++ m2-capstone/src/main/java/DAOInterfacesAndJavaBeans/SiteDAO.java
@@ -8,5 +8,5 @@ public interface SiteDAO {
 	
 	public List<Site>getSitesByID();
 	
-	public List<Site> getAvailSites(String from_date, String to_date, int campground_id);
+	public List<Site> getAvailSites(String from_date, String to_date, long campground_id);
 }
